Override gitlab_backup to gitswarm_backup
The text isn't easily replaceable in regexes and there are uses of it in Dir.glob. This resulted in my having to copy and override more methods than I wanted, and not being able to call super.

Going forward we should try and push a change back to community that would make the backup name confirgurable. Then we could just override IO to change some of the output messages.