Added 'finish' logic for initial import. Runs via sidekiq and every 5 min.
The sidekiq event is our intended handler and should promptly end import mode.
If for some reason it doesn't run (e.g. the system is restarted mid-import)
the MirrorFetch worker (which runs every 5 minutes) will call finish when it
detects the repo has completed a fetch but is flagged as importing.

This won't work till !202 goes in though.