Fix users access to non-authorized projects when enforcing_permissions
Just because a project isn't in the list of a user's authorized projects, doesn't mean they don't have access. Public/internal projects in gitswarm should be accesible if the user can see them on the perforce side.

For this change I moved the repo access caching into it's own class to make calling it from various places easier.

Note: At the moment on login you are going to clear the user's repo access cache, and repopulate their authorized_projects, but you may not be repopulating the cache for these public/internal projects. For now at least, until we get in some better cache handling.