Move the p4_tree coffeescript into it's own file
Add some more tree config to give it the selection style that we want for this first go. (only select one path, don't cascade selections down.

Also added some very basic save mapping functionality. The mapping makes it to the create_repository method in the project model at the moment, and then ends in a TODO.