Sanitize branch input from the custom repo creator.
In the projects controller we strip tags on the branchnames. Which matches GitLab's behaviour in the branch create action.
In the javascript, we html escape the branch input when inserting it into the dom as for display in the places we weren't already. (some places were using jquery's .text and were fine.
In the javascript, we also use underscore.js's escape method to make the branch name safe to put into html attributes. (quotes, and a few other special characters get escaped)

We now use document.getElementsByName when wanting to search for the branch input, instead of jquery's sizzle selectors, as they suffered from sometimes needing quote escapement and sometimes not. (If they needed it for " they wouldn't need it for ' and would break)